[FIX] App modularization, minimum unittest
