+++ Pipfile
@@ -7,6 +7,7 @@ name = "pypi"
 flask = "*"
 
 [dev-packages]
+pylint = "*"
 
 [requires]
 python_version = "3.6"
