+++ app/users.py
@@ -1,0 +1,5 @@
+"""This module manages the users"""
+
+class User(object):
+    """This class manages the app users"""
+    pass
