+++ requirements.txt
@@ -1,7 +1,14 @@
+astroid==1.5.3
 click==6.7
+colorama==0.3.9
 Flask==0.12.2
+isort==4.2.15
 itsdangerous==0.24
 Jinja2==2.9.6
+lazy-object-proxy==1.3.1
 MarkupSafe==1.0
+mccabe==0.6.1
+pylint==1.7.4
+six==1.11.0
 Werkzeug==0.12.2
-Flask-Testing==0.6.2
+wrapt==1.10.11
