+++ run.py
@@ -1,5 +1,5 @@
 """Runs the dev server"""
-from app.app import APP
+from app import APP
 
 if __name__ == "__main__":
     APP.run()
